+++ packages/plugin/src/Fields/FieldInterface.php
@@ -49,6 +49,7 @@ interface FieldInterface
     public const TYPE_CALCULATION = 'calculation';
     public const TYPE_RADIO_GROUP = 'radios';
     public const TYPE_HTML = 'html';
+    public const TYPE_SUMMARY = 'summary';
     public const TYPE_SUBMIT = 'submit';
     public const TYPE_SAVE = 'save';
     public const TYPE_DYNAMIC_RECIPIENTS = 'dynamic-recipients';